pyproject.toml
[tool.poetry]
name = "cloudfront-test"
version = "0.1.0"
description = ""
authors = ["Alexander Dich Jensen <dichjensen@gmail.com>"]
readme = "README.md"
packages = [{include = "cloudfront_test"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
construct = "^2.10.70"
aws-cdk-lib = "^2.141.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
